Fix for problem moving pages around in an exhibit
This behavior was noticed after an exhibit had a significant number
of pages.  Symptoms: It was impossible to move the last page upward,
or it was impossible to move one of the top items downward past
a certain point, etc.
The following post gave me a hint:
http://omeka.org/forums/topic/item-file-re-ordering
It references the following commit on github.com/omeka/Omeka:
3306d999e4113984508ed210cc0653a896709ad8
So the change here is based on those changes. Once the change was
made, the problem was resolved.
Note that the posting mentions a bug in jQuery. Not sure if that is
actually the case, but if so, and the bug is fixed, this change may
need to be undone / rolled back.